chore: extract gcs client from GCSPath
 - Add `BucketClient` class that can be implemented by any client that supports read/write bucket APIs.
 - Add `BucketClientGCS` for the GCS service
