+++ gcspath/__init__.py
@@ -1,2 +1,2 @@
-from .gcs import GCSPath, PureGCSPath  # noqa
+from .api import GCSPath, PureGCSPath  # noqa
 from .client import BucketDirEntry, BucketStatResult  # noqa
